--- nadro-credit/handler.ts
@@ -1,5 +1,0 @@
-import { SendEmailNotificationController } from './infrastructure/controllers/SendEmailNotificationController';
-
-export const login = async (event: any) => {
-  return await SendEmailNotificationController.handle(event);
-};
