+++ ppi-services/domain/entities/CreditBehavior.ts
