+++ ppi-services/domain/types/CustomerDocs.type.ts
