+++ ppi-services/handler.ts
@@ -1,0 +1,5 @@
+import { LoginPPIController } from './infrastructure/controllers/PPILoginController';
+
+export const login = async (event: any) => {
+  return await LoginPPIController.handle(event);
+};
